+++ stories/components/Input.stories.tsx
@@ -21,7 +21,10 @@ export const InputStory = () => (
         <div className="story-info">
             <Alert tone="critical">
                 <strong>Deprecated:</strong> Please use{' '}
-                <a href="reactist/?path=/docs/design-system-textfield">TextField</a> instead
+                <a href={`${window.location.origin}?path=/docs/design-system-textfield`}>
+                    TextField
+                </a>{' '}
+                instead
             </Alert>
             <p>
                 This component is a dumb wrapper around the